+++ iti-react/Changelog.md
@@ -79,3 +79,6 @@ Then:
 
 - `Select` components: only search the `label` when filtering
 
+# [wip]
+
+- Core changes